adapt kernel playground for IOAM6 Global-Opaque Block development
The public kernel-playground repository has been modified to support
private IOAM6 kernel and userspace development.

In particular, the Global-Opaque Block (GOB) has been introduced and the
kernel now allows a custom eBPF program to be attached to each block.
The program is executed Hop-by-Hop and can read and write within
the GOB space that is embedded in the packet.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>