set gobschema on athos, porthos and aramis
The script configures the gob schema on the nodes athos, porthos and
aramis using some samples eBPF programs that will run hop by hop.

Run the script as root to build the namespaces, load the eBPF objects
and start the tmux UI.
In the tmux window named athos execute:

  ping db22::2

to generate traffic that traverses the IOAM-6 and SRv6 hops.

Switch to the beta window (or ip netns exec beta bash) and read
/sys/kernel/tracing/trace_pipe to see the eBPF messages emitted at each
hop.
Detach from tmux with Ctrl-b d or exit the script to trigger cleanup of
namespaces, veth devices and the tmux session.

Signed-off-by: Andrea Mayer <andrea.mayer@uniroma2.it>